requirements.txt
requests==2.31.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
tenacity==8.2.3
python-dateutil==2.8.2
python-dotenv==1.0.0
